Updated: Editor -> AdjustZoom int to double para for conversion with float

+++ HousecatEngine/src/editor/Editor.cpp
@@ -281,7 +281,7 @@ void Editor::KeyboardCameraController(SDL_Event& event) {
 	}
 }
 
-void Editor::AdjustZoom(int direction) {
+void Editor::AdjustZoom(double direction) {
 	constexpr float zoomStep = 0.1f;
 
 	if (direction > 0) {
